fix: fail workflow on subflow prep errors
