+++ tests/unit/test_dsl_workflow_concurrency.py
@@ -8,12 +8,14 @@ from typing import Any, cast
 from unittest.mock import AsyncMock, patch
 
 import pytest
-from temporalio.exceptions import ApplicationError
+from temporalio.exceptions import ActivityError, ApplicationError
 
 from tracecat import config
 from tracecat.auth.types import Role
+from tracecat.dsl.action import DSLActivities
 from tracecat.dsl.common import DSLEntrypoint, DSLInput, DSLRunArgs
-from tracecat.dsl.enums import FailStrategy, WaitStrategy
+from tracecat.dsl.enums import FailStrategy, PlatformAction, WaitStrategy
+from tracecat.dsl.scheduler import DSLScheduler
 from tracecat.dsl.schemas import (
     ROOT_STREAM,
     ActionRetryPolicy,
@@ -73,7 +75,27 @@ def _build_workflow(*, limits: EffectiveLimits | None = None) -> DSLWorkflow:
     context = ExecutionContext(ACTIONS={}, TRIGGER=None)
     workflow.context = context
     workflow.scheduler = cast(Any, SimpleNamespace(streams={ROOT_STREAM: context}))
+    assert workflow.role.workspace_id is not None
+    workflow.workspace_id = workflow.role.workspace_id
+    workflow.wf_exec_id = workflow.run_context.wf_exec_id
     return workflow
+
+
+def _activity_error_from(
+    cause: Exception, *, activity_type: str = "prepare_subflow_activity"
+) -> ActivityError:
+    try:
+        raise ActivityError(
+            "Activity failed",
+            scheduled_event_id=1,
+            started_event_id=2,
+            identity="test",
+            activity_type=activity_type,
+            activity_id=activity_type,
+            retry_state=None,
+        ) from cause
+    except ActivityError as e:
+        return e
 
 
 @pytest.mark.anyio
@@ -236,6 +258,75 @@ async def test_execute_task_releases_action_permit_when_cancelled_during_heartbe
 
     assert result.get_data() == {"ok": True}
     release_mock.assert_awaited_once_with(action_id="permit-id")
+
+
+@pytest.mark.anyio
+async def test_prepare_subflow_activity_failure_in_scatter_fails_workflow() -> None:
+    workflow = _build_workflow()
+    dsl = DSLInput(
+        title="test",
+        description="test",
+        entrypoint=DSLEntrypoint(ref="scatter"),
+        actions=[
+            ActionStatement(
+                ref="scatter",
+                action=PlatformAction.TRANSFORM_SCATTER,
+                args={"collection": ["item"]},
+            ),
+            ActionStatement(
+                ref="call_child",
+                action=PlatformAction.CHILD_WORKFLOW_EXECUTE,
+                args={"workflow_alias": "child"},
+                depends_on=["scatter"],
+            ),
+            ActionStatement(
+                ref="handle_error",
+                action="core.noop",
+                depends_on=["call_child.error"],
+            ),
+        ],
+    )
+    scheduler = DSLScheduler(
+        executor=workflow.execute_task,
+        dsl=dsl,
+        max_pending_tasks=16,
+        context=ExecutionContext(ACTIONS={}, TRIGGER=None),
+        role=workflow.role,
+        run_context=workflow.run_context,
+    )
+    workflow.dsl = dsl
+    workflow.scheduler = scheduler
+    executed_refs: list[str] = []
+
+    async def execute_activity(activity: object, *_: Any, **__: Any) -> object:
+        if activity == DSLActivities.handle_scatter_input_activity:
+            return InlineObject(data=["item"], typename="list")
+        if activity == DSLActivities.prepare_subflow_activity:
+            raise _activity_error_from(RuntimeError("prepare failed"))
+        raise AssertionError(f"Unexpected activity: {activity}")
+
+    async def run_action(task: ActionStatement) -> InlineObject:
+        executed_refs.append(task.ref)
+        return InlineObject(data={"handled": True})
+
+    with (
+        patch(
+            "tracecat.dsl.scheduler.workflow.execute_activity",
+            new=AsyncMock(side_effect=execute_activity),
+        ),
+        patch(
+            "tracecat.dsl.workflow.workflow.execute_activity",
+            new=AsyncMock(side_effect=execute_activity),
+        ),
+        patch.object(workflow, "_run_action", new=AsyncMock(side_effect=run_action)),
+    ):
+        task_exceptions = await scheduler.start()
+
+    assert task_exceptions is not None
+    assert "call_child" in task_exceptions
+    assert "prepare failed" in task_exceptions["call_child"].details.message
+    assert executed_refs == []
+    assert not scheduler.stream_exceptions
 
 
 @pytest.mark.anyio
