+++ tracecat/dsl/action.py
@@ -67,9 +67,6 @@ from tracecat.validation.schemas import ValidationDetail
 
 _thread_local = threading.local()
 
-MATERIALIZE_CONTEXT_ERROR_MESSAGE = "Failed to materialize context"
-PLATFORM_EXECUTION_ERROR_TYPE = "PlatformExecutionError"
-
 
 def _strip_string_values(args: dict[str, Any]) -> dict[str, Any]:
     """Strip leading/trailing whitespace from string values in args.
@@ -346,9 +343,8 @@ async def materialize_context(ctx: ExecutionContext) -> MaterializedExecutionCon
     except Exception as e:
         logger.warning("Error materializing context", error=e)
         raise ApplicationError(
-            MATERIALIZE_CONTEXT_ERROR_MESSAGE,
+            "Failed to materialize context",
             non_retryable=True,
-            type=PLATFORM_EXECUTION_ERROR_TYPE,
         ) from e
 
     # Reconstruct ACTIONS dict with materialized results
